+++ src/main/php/lang/ast/emit/PHP81.class.php
@@ -21,6 +21,7 @@ use lang\ast\types\{
 class PHP81 extends PHP {
   use
     EmulatePipelines,
+    RemoveVoidCasts,
     RewriteBlockLambdaExpressions,
     RewriteCallables,
     RewriteCloneWith,