+++ src/main/php/lang/ast/emit/PHP83.class.php
@@ -21,6 +21,7 @@ use lang\ast\types\{
 class PHP83 extends PHP {
   use 
     EmulatePipelines,
+    RemoveVoidCasts,
     RewriteCallables,
     RewriteCloneWith,
     RewriteBlockLambdaExpressions,