+++ src/main/php/lang/ast/emit/PHP84.class.php
@@ -19,12 +19,7 @@ use lang\ast\types\{
  * @see  https://wiki.php.net/rfc#php_84
  */
 class PHP84 extends PHP {
-  use
-    EmulatePipelines,
-    RewriteCallables,
-    RewriteCloneWith,
-    RewriteBlockLambdaExpressions
-  ;
+  use EmulatePipelines, RemoveVoidCasts, RewriteCallables, RewriteCloneWith, RewriteBlockLambdaExpressions;
 
   public $targetVersion= 80400;